c4PeerDiscovery.hh no longer uses nonpublic headers
